+++ source/Popup/index.tsx
@@ -5,11 +5,9 @@ import ReactDOM from 'react-dom';
 // Common styles
 import '../styles/main.scss';
 
-import './styles.scss';
-
 import Popup from './Popup';
 
-// eslint-disable-next-line import/no-webpack-loader-syntax, import/no-unresolved,  @typescript-eslint/no-var-requires
+// eslint-disable-next-line import/no-webpack-loader-syntax, import/no-unresolved,  @typescript-eslint/no-var-requires, node/no-missing-require
 const theme = require('sass-extract-loader?{"plugins": ["sass-extract-js"]}!../styles/base/_variables.scss');
 // Require sass variables using sass-extract-loader and specify the plugin
