+++ source/Popup/refactor/styles.scss
@@ -1,7 +1,7 @@
-@import "../styles/base/fonts";
-@import "../styles/base/reset";
-@import "../styles/base/variables";
-@import "../styles/base/components";
+@import "../../styles/base/fonts";
+@import "../../styles/base/reset";
+@import "../../styles/base/variables";
+@import "../../styles/base/components";
 
 body {
   color: $black;
